Updated structure and introduced _reference key
Changed the design to no longer use keyed objects for the resources as this wont guarantee order of execution whereas an array of objects does guarantee order. This then needed a new key on the object for tracking the identifier in reference so the `_reference` key was introduced for this purpose.